refactor: replace error classes with super-class and enums, rename model files
